requirements.txt
aioredis==2.0.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.35.27
botocore==1.35.27
certifi==2024.8.30
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
dj-database-url==2.2.0
Django==4.2.16
django-compat==1.0.15
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-guardian==2.4.0
django-rest-passwordreset==1.4.1
django-storages==1.14.4
django4-background-tasks==1.2.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
environ==1.0
gunicorn==23.0.0
hiredis==2.3.2
hyperlink==21.0.0
idna==3.7
importlib_metadata==7.1.0
incremental==22.10.0
jmespath==1.0.1
Markdown==3.6
msgpack==1.0.8
packaging==24.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyHamcrest==2.1.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
requests==2.32.3
s3transfer==0.10.2
service-identity==24.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
stripe==11.1.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
urllib3==1.26.20
whitenoise==6.7.0
zipp==3.18.1
zope.interface==6.3
